Adding sprite demo

+++ src/Model/Widget/FloatPosition.php
@@ -26,7 +26,7 @@ final class FloatPosition implements Stringable
         return (false === $xBounds->contains($this->x)) || (false === $yBounds->contains($this->y));
     }
 
-    public function update(int $x, int $y): void
+    public function update(float $x, float $y): void
     {
         $this->x = $x;
         $this->y = $y;
